Rename model to entities package
